requirements.txt
psycopg2-binary==2.9.7
pandas==2.0.3
numpy==1.24.3
tqdm==4.66.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0
sqlalchemy==2.0.23
asyncpg==0.29.0
jellyfish==1.0.1
textdistance==4.5.0
